agx_driver_api/gripper_sdk/pyproject.toml
[build-system]
# requires = ["setuptools >= 40.0", "setuptools_scm>=8"]
requires = ["setuptools >= 40.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agx_motor_api"
dynamic = ["readme", "version"]
description = "A sdk to control motor driver"
authors = [
    {name = "RosenYin", email = "yinruocheng321@gmail.com"},
]
license = {text = "MIT License"}
requires-python = ">=3.6"
dependencies = [
    "python-can>=3.3.4",
]
classifiers = [
    "Operating System :: POSIX :: Linux",
]

[tool.setuptools.dynamic]
readme = { file = "README.MD" }